complete

+++ Admission/nghean2024-2025_10/SODEP.cpp
@@ -5,22 +5,8 @@
 using namespace std;
 const ll inf=LLONG_MAX;
 const ll mod=1e9+7;
-int n;
-int l,r;
-// void zero()
-// {
-//     for(int i=1;i<=1e7;i++)
-//     {
-//         // int dem=0;
-//         int m=i;
-//         while(m>0)
-//         {
-//             a[i]+=m/5;
-//             m/=5;
-//         }
-//         // a[i]=a[i-1]+a[i];
-//     }
-// }
+ll n;
+ll l,r;
 kien()
 {
     // if(fopen(".inp","r"))
@@ -31,14 +17,14 @@ kien()
     ios_base::sync_with_stdio(0);
     cin.tie(0);
     cout.tie(0);
-    int t;
+    ll t;
     cin>>t;
     // zero();
     while(t--)
     {
         cin>>l>>r;
         l--;
-        int a=0,b=0;
+        ll a=0,b=0;
         while(l>0)
         {
             a+=l/5;
